owr: Change MediaSource API to allow embedding in other pipelines
Instead of a get_pad() there is now a get_source_element() which will
return a GStreamer source element that can be added to any bin and
will provide the requested data.

As a side effect this also makes state changes more sensible and
removes the need of a valve in the media sources.